backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
google-generativeai>=0.8.3
supabase==2.3.1
python-dotenv==1.0.1
pydantic==2.6.3
thefuzz==0.22.1
python-Levenshtein==0.25.0
